# Conflicts:
#	Progra3/.vs/Progra3/v15/.suo

+++ Progra3/App_Code/Product.cs
@@ -7,11 +7,11 @@ public class Product
 {
     public int id { get; set; }
     public String descripcion { get; set; }
-    public double precio { get; set; }
+    public int precio { get; set; }
     public int tiempoRealizacion { get; set; }
     public int cantidad { get; set; }
 
-    public Product (int id, String descripcion, double precio, int tiempoRealizacion , int cantidad)
+    public Product (int id, String descripcion, int precio, int tiempoRealizacion , int cantidad)
     {
         this.id = id;
         this.descripcion = descripcion;
